requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
anthropic==0.28.0
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
msal==1.28.0
aiofiles==23.2.1
Pillow==10.3.0
pdf2image==1.17.0
pypdf==4.2.0
